Refactored free_test.rs -> flight.rs
